Refactor fetchWithProgress
